requirements.txt
line-bot-sdk==1.16.0
flask==1.1.2
pandas==1.0.3
numpy==1.18.2
requests==2.23.0
beautifulsoup4==4.9.0